chore: users can only sync when connected to decred network chore: add button to manually sync proposal incase initial sync fails chore: implemented changes requested from PR review
