+++ app/src/main/res/values/strings.xml
@@ -587,5 +587,13 @@
     <string name="new_proposal">New Proposal</string>
     <string name="vote_started">Vote Started</string>
     <string name="vote_ended">Vote Ended</string>
+    <string name="politeia_server_url" translatable="false">http://proposals.decred.org/proposals/</string>
+    <string name="share_proposal">Share Proposal Link</string>
+    <string name="text_pain">text/plain</string>
+    <string name="comments">%d Comments</string>
+    <string name="version_number">version %s</string>
+    <string name="proposal_percentage">%.2f%%</string>
+    <string name="syncing_proposals">Syncing Proposals</string>
+    <string name="error_syncying_proposals">There was an error syncing proposals, visit the proposal page to try again</string>
     <!--/Politeia-->
 </resources>
